refactor : movieCard 클래스 명에 맞게 파일명 수정
